client push

+++ src/components/Todo.js
@@ -1,6 +1,8 @@
 import React from "react";
 import styles from "../styles/Todos.module.css";
 import { AuthConsumer } from "../store/auth.js";
+import { useNavigate } from "react-router-dom";
+import {toast} from "react-toastify";
 
 function Todo({
   item,
@@ -9,20 +11,30 @@ function Todo({
   setTodo,
   setEditId,
   setOriginalTodo,
+  setIsUpdateMode,
 }) {
+  const navigate = useNavigate();
   const { token } = AuthConsumer();
 
   const handleDeleteButton = async (id) => {
+    const isConfirmed = window.confirm("Do you want to Delete this Todo");
+    if(!isConfirmed){
+      return;
+    }
     const response = await fetch(`http://localhost:3001/todo/delete/${id}`, {
       method: "DELETE",
       headers: {
-        Authorization: `Bearer ${token}`,
+        'Authorization': `Bearer ${token}`,
       },
     });
     const res_data = await response.json();
     if (response.status === 200) {
       console.log("todo is deleted");
       setShouldRefresh((prev) => !prev);
+    }else if(response.status === 401){
+      toast.error(res_data.message);
+    }else if(response.status === 404){
+      toast.error(res_data.message);
     }
   };
   console.log("item", item);
@@ -31,13 +43,15 @@ function Todo({
     if (e.target.value === "update") {
       handleUpdateOption(item);
     } else if (e.target.value === "updateOnOtherPage") {
-      handleUpdateOnOtherPageOption(item);
+      // handleUpdateOnOtherPageOption(item);
+      navigate(`/edit/${item.id}`);
+
     }
     e.target.value = "";
   };
-  const handleUpdateOnOtherPageOption = (item) => {
+  // const handleUpdateOnOtherPageOption = (item) => {
     
-  };
+  // };
   const handleUpdateOption = async (item) => {
     console.log("in update btn handler");
     const todoItem = {
@@ -45,9 +59,13 @@ function Todo({
       description: item.description,
       date: item.date,
     };
+
+    console.log("in todo todoItem for update", todoItem);
+    
     setTodo(todoItem);
     setOriginalTodo(todoItem);
     setEditId(item.id);
+     setIsUpdateMode(true);
   };
   return (
     <>
@@ -55,7 +73,8 @@ function Todo({
         <div className={styles.todoDetails}>
           <p>{item.title}</p>
           <p>{item.description}</p>
-          <p>{new Date(item.date).toLocaleDateString()}</p>
+          <p>{item.date}</p>
+          {/* <p>{new Date(item.date).toLocaleDateString()}</p> */}
         </div>
 
         <div className={styles.todoActions}>
@@ -84,17 +103,3 @@ function Todo({
 
 export default Todo;
 
-//   <li className={styles.todo}>
-//     <p>{item.title}</p>
-//     <p>{item.description}</p>
-//     <p>{new Date(item.date).toLocaleDateString()}</p>
-
-//   <select>
-//     <option>Update</option>
-//     <option>Update On other page</option>
-//   </select>
-//   <select>
-//     <option>Delete</option>
-//     <option>Delete On other page</option>
-//   </select>
-//   </li>
