+++ src/pages/DeleteProfile.js
@@ -1,16 +1,21 @@
-import React, { useEffect } from 'react'
+import React, { useEffect, useRef } from 'react'
 import { AuthConsumer } from '../store/auth'
 import {toast} from "react-toastify";
-import { redirect, useNavigate } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 
 function DeleteProfile() {
-    let called = false;
-    const {token, setToken, setIsLoggedIn} = AuthConsumer();
+    let called = useRef(false);
+    const {token, setToken, setIsLoggedIn, setUserprofilePicture} = AuthConsumer();
     const navigate = useNavigate();
-     useEffect(() => {
-        if(called) return;
-        called = true;
+    useEffect(() => {
+        if (called.current) return;
+        called.current = true;
         const deleteUser = async () => {
+            const isConfirmed = window.confirm("Do you want to Delete Profile");
+            if (!isConfirmed) {
+                navigate(-1)
+                return;
+            }
             const response = await fetch("http://localhost:3001/user/delete", {
                 method: "DELETE",
                 headers:{
@@ -22,10 +27,15 @@ function DeleteProfile() {
             console.log("In delete user page",res_data);
             
             if(response.status === 200){
+                console.log("Top in delete success");
+                
                 setToken("");
+                setUserprofilePicture("");
                 setIsLoggedIn(false);
                 toast.success(res_data.message);
+                
                 navigate("/register");
+                console.log("Bottom in delete success");
             }else if(response.status === 404){
                 toast.error(res_data.message);
             }else if(response.status === 401){
@@ -39,7 +49,7 @@ function DeleteProfile() {
             }
         }
         deleteUser();
-    }, [])
+    }, [token, setIsLoggedIn, setToken, setUserprofilePicture, navigate])
   return (
    <></> 
   )
